day3: completed
The level of difficulty fooled me. I thought it was easy but I spent
01:23:23.59 (1hr and 24mins).

The serialization was a piece of cake but the deserialization was hard.
I made a simple mistake and spent 30 mins to recover my mistake.

I was about to give up when it gets to 1hr yet I finished it. No matter
how long it took me.